refactor(frontend): move API layer from other module to book module, move api base URL to env
