+++ frontend/src/api/book/book.api.ts
@@ -8,9 +8,23 @@ import type {
     BooksListParams, 
     BooksListResponseDto, 
     RemoveBookDto, 
-    RemoveBookResponseDto 
+    RemoveBookResponseDto,
+    AddAuthorDto, 
+    AddAuthorResponseDto, 
+    AddCopyDto,
+    AddCopyResponseDto, 
+    AddPublisherDto, 
+    AddPublisherResponseDto, 
+    ApiRemoveCopyResponseDto, 
+    RemoveCopyDto 
 } from "./book.types";
-import { mapAdd, mapBookDataResponse, mapBooksListResponse } from "./book.mapper";
+import { 
+    mapAdd, 
+    mapBookDataResponse, 
+    mapBooksListResponse,
+    mapAddAuthorResponse, 
+    mapAddPublisherResponse
+ } from "./book.mapper";
 
 const baseURL = "/book";
 
@@ -33,5 +47,30 @@ export const bookApi = {
     bookData: async (id: string): Promise<BookData> => {
         const response = await api.get<BookDataResponseDto>(`${baseURL}/${id}`);
         return mapBookDataResponse(response);
+    },
+
+    addPublisher: async (data: AddPublisherDto): Promise<{ id: string, publisherName: string }> => {
+        const response = await api.post<AddPublisherResponseDto>(`/author/addAuthor`, data);
+        return mapAddPublisherResponse(response);
+    },
+
+    copyAdd: async (data: AddCopyDto): Promise<{ id: string }> => {
+        const response = await api.post<AddCopyResponseDto>(`/copy/add`, data);
+        return { id: response.id };
+    },
+
+    copyRemove: async (data: RemoveCopyDto): Promise<string> => {
+        const response = await api.patch<ApiRemoveCopyResponseDto>(`/copy/remove`, data);
+        return response.message;
+    },
+
+    addAuthor: async (data: AddAuthorDto): Promise<{ id: string, authorName: string }> => {
+        const response = await api.post<AddAuthorResponseDto>(`/publisher/addPublisher`, data);
+        return mapAddAuthorResponse(response);
+    },
+
+    healthStatus: async (): Promise<string> => {
+        const response = await api.get<{ status: string }>(`/health`);
+        return response.status;
     },
 };
