+++ frontend/src/api/book/book.mapper.ts
@@ -1,5 +1,7 @@
 import type { 
+    AddAuthorResponseDto,
     AddBookResponseDto, 
+    AddPublisherResponseDto, 
     Book, BookData, 
     BookDataResponseDto,
     BookDto,
@@ -26,4 +28,14 @@ export const mapBookDataResponse = (dto: BookDataResponseDto): BookData => ({
     publisherId: dto.publisher_id,
     ISBN: dto.ISBN,
     numberOfCopies: dto.number_of_copies
+});
+
+export const mapAddPublisherResponse = (dto: AddPublisherResponseDto): { id: string, publisherName: string } => ({
+    id: dto.id,
+    publisherName: dto.publisher_name
+});
+
+export const mapAddAuthorResponse = (dto: AddAuthorResponseDto): { id: string, authorName: string } => ({
+    id: dto.id,
+    authorName: dto.author_name
 });
