+++ frontend/src/api/book/book.types.ts
@@ -16,6 +16,22 @@ export interface BooksListParams {
     search?: string;
 }
 
+export interface AddPublisherDto {
+    publisher_name: string;
+}
+
+export interface AddCopyDto {
+    book_id: string;
+}
+
+export interface RemoveCopyDto {
+    id: string;
+}
+
+export interface AddAuthorDto {
+    author_name: string;
+}
+
 //Response (only data)
 export interface AddBookResponseDto {
     id: string;
@@ -41,6 +57,25 @@ export interface BookDataResponseDto {
     publisher_id: string;
     ISBN: string;
     number_of_copies: number;
+}
+
+export interface AddPublisherResponseDto {
+    id: string;
+    publisher_name: string;
+}
+
+export interface AddCopyResponseDto {
+    id: string;
+}
+
+export interface AddAuthorResponseDto {
+    id: string;
+    author_name: string;
+}
+
+//Response (data + messenge + itp.)
+export interface ApiRemoveCopyResponseDto {
+    message: string;
 }
 
 //interface model
