+++ frontend/src/api/client.ts
@@ -1,4 +1,4 @@
-const BASE_URL = "http://localhost:8080/api";
+const BASE_URL = import.meta.env.VITE_API_BASE_URL;
 
 export const client = async <T>(
   url: string,
